backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
Flask-Cors
